Extracted reversed list class to upper level
